# feature: improve stable stringify