This is a performance improvement in many cases. For big / deep
objects it will be a minor performance penalty in the case no
options are used.
If options are used (i.e. replacer or spacer) it will always be
faster.

The implementation is side effect free, so it now works 1-to-1 as
JSON.stringify.

For performance reasons this is a lot of duplicated code... I do not see a much nicer implementation right now that could do the same though :/.

No options:
```
old:   simple object x 916,356 ops/sec ±0.95% (95 runs sampled)
old:   circular      x 367,928 ops/sec ±1.00% (94 runs sampled)
old:   deep          x 16,704 ops/sec ±1.01% (91 runs sampled)
old:   deep circular x 16,234 ops/sec ±1.44% (91 runs sampled)

fast-safe-stringify:   simple object x 1,389,123 ops/sec ±0.63% (96 runs sampled)
fast-safe-stringify:   circular      x 477,570 ops/sec ±1.05% (90 runs sampled)
fast-safe-stringify:   deep          x 15,363 ops/sec ±1.00% (91 runs sampled)
fast-safe-stringify:   deep circular x 15,175 ops/sec ±0.90% (95 runs sampled)
```

Spacer
```
old:   simple object x 966,154 ops/sec ±1.09% (91 runs sampled)
old:   circular      x 295,570 ops/sec ±0.55% (95 runs sampled)
old:   deep          x 13,203 ops/sec ±0.81% (95 runs sampled)
old:   deep circular x 12,707 ops/sec ±1.10% (94 runs sampled)

fast-safe-stringify:   simple object x 1,214,236 ops/sec ±1.08% (93 runs sampled)
fast-safe-stringify:   circular      x 428,453 ops/sec ±0.82% (88 runs sampled)
fast-safe-stringify:   deep          x 14,071 ops/sec ±0.78% (92 runs sampled)
fast-safe-stringify:   deep circular x 13,856 ops/sec ±0.98% (92 runs sampled)
```

Replacer
```
old:   simple object x 939,815 ops/sec ±2.50% (90 runs sampled)
old:   circular      x 210,840 ops/sec ±0.72% (91 runs sampled)
old:   deep          x 12,708 ops/sec ±0.98% (91 runs sampled)
old:   deep circular x 12,358 ops/sec ±0.81% (90 runs sampled)

fast-safe-stringify:   simple object x 1,166,188 ops/sec ±0.71% (92 runs sampled)
fast-safe-stringify:   circular      x 445,552 ops/sec ±1.10% (92 runs sampled)
fast-safe-stringify:   deep          x 14,685 ops/sec ±1.04% (95 runs sampled)
fast-safe-stringify:   deep circular x 14,540 ops/sec ±0.89% (94 runs sampled)
```

Both
```
old:   simple object x 957,466 ops/sec ±1.20% (89 runs sampled)
old:   circular      x 194,075 ops/sec ±0.66% (93 runs sampled)
old:   deep          x 10,923 ops/sec ±0.86% (89 runs sampled)
old:   deep circular x 10,514 ops/sec ±1.26% (92 runs sampled)

fast-safe-stringify:   simple object x 1,117,194 ops/sec ±0.94% (92 runs sampled)
fast-safe-stringify:   circular      x 407,977 ops/sec ±1.06% (90 runs sampled)
fast-safe-stringify:   deep          x 13,700 ops/sec ±1.26% (93 runs sampled)
fast-safe-stringify:   deep circular x 13,570 ops/sec ±1.18% (95 runs sampled)
```