@davidmarkclements 2018-01-13
so... from what I can see you've reimplemented `JSON.stringify`

First to be clear, let me just say: awesome. 

Before we can go forward though, we need to benchmark this excessively for as many scenarios as possible vs JSON.stringify 

We know from past experiments that getting on par with `JSON.stringify` was 
pretty much impossible because of the TCO trick they use in 
the native implementation – so we need to be extremely thorough in establishing this is a performance improvement across the board

What we don't want is people using fast-safe-stringify having a noticeable slow-down in their use cases

@mcollina what are your thoughts?