@BridgeAR 2018-01-13
I understand the concerns about it. My main reason was to see how well it performs and i wanted to overcome the negative implementation parts.

I only implemented this for the stable version because the regular one is faster the way it is right now. With the added overhead in the stable version it makes sense to have this implementation though.

What we can do is to add lots of tests to make sure it works fine.