@mcollina 2018-01-13
How about we do: fast-stable-stringify? I prefer small modules with a single purpose.