@BridgeAR 2018-01-13
The original intention against it was that there are multiple out there (all with "fast" in their name...) but I will go ahead and create a new package when I am at home. We can reference to it :)